Fix warning in init_rs.h

+++ src/libs/fec/init_rs.h
@@ -8,7 +8,7 @@
 
   rs = NULL;
   /* Check parameter ranges */
-  if(symsize < 0 || symsize > 8*sizeof(data_t)){
+  if(symsize < 0 || (size_t)symsize > 8*sizeof(data_t)){
     goto done;
   }
